fix(m4): harden RTF/ODT/EPUB extractors per 2nd review (#77)
Address findings from a second 6-axis correctness/robustness review:

- ODT: `<text:s text:c="0"/>` now yields zero spaces (was incorrectly 1);
  large counts still clamp to 32.
- Shared walker: a self-closing block element (`<text:p/>`, `<p/>`) now
  advances the per-element anchor counter so node_path positions stay
  stable regardless of `<text:p/>` vs `<text:p></text:p>` serialization.
- EPUB: classify on the XML local name so namespace-prefixed XHTML
  (`<x:p>`, `<x:h1>`) is recognised instead of silently dropping a chapter.
- EPUB: per-spine-entry decompression-bomb guard now runs BEFORE walking
  the entry (tighter peak-memory bound), in addition to the cumulative cap.
- SourceAnchor: `Rtf.text_offset` and `Epub.spine_index` are now `u64`
  (not `usize`) for a target-independent JSON wire shape (matches Pdf).
- Cargo.toml: document the rbook → zip 8.6.0 → flate2 zlib-rs backend
  feature-unification side effect.
- Doc-comment accuracy (text:s clamp behaviour; EPUB lists/tables).

Tests: text:c zero/clamp, self-closing-counter node_path, multibyte
through the shared walker, EPUB namespaced elements, and a full-pipeline
RTF end-to-end ingest (extract → chunk → embed → index) via the real
extractor. All extractor + ingest suites green.

+++ lens-core/src/extract/mod.rs
@@ -68,16 +68,18 @@ pub enum SourceAnchor {
     /// An RTF block: its byte offset in the canonical extracted-text buffer.
     ///
     /// Additive under its own `kind` tag (issue #77); cannot change the wire
-    /// shape of any pre-existing variant.
-    Rtf { text_offset: usize },
+    /// shape of any pre-existing variant. `u64` (not `usize`) so the persisted
+    /// JSON wire shape is identical on 32- and 64-bit targets.
+    Rtf { text_offset: u64 },
     /// An ODT block: the node path in `content.xml` (e.g. `"body/text:h[0]"`).
     ///
     /// Additive under its own `kind` tag (issue #77).
     Odt { node_path: String },
     /// An EPUB block: the spine index and the href of the content document.
     ///
-    /// Additive under its own `kind` tag (issue #77).
-    Epub { spine_index: usize, href: String },
+    /// Additive under its own `kind` tag (issue #77). `spine_index` is `u64`
+    /// (not `usize`) for a target-independent JSON wire shape.
+    Epub { spine_index: u64, href: String },
 }
 
 /// The canonical extraction result for a single source.
